requirements.txt
requests
beautifulsoup4
transformers
fpdf
yagmail
schedule
python-dotenv
torch